ver 0.5.10

+++ IL2DCE/IL2DCE/MissionObjectModel/SilkySkyCloDFile.cs
@@ -565,7 +565,7 @@ namespace IL2DCE.MissionObjectModel
             }
         }
 
-        public static void DeleteSectionAndSubSection(ISectionFile fileSrc, string section, string [] SubSectionAddKey)
+        public static void DeleteSectionAndSubSection(ISectionFile fileSrc, string section, string [] SubSectionAddKey = null)
         {
             int lines = fileSrc.lines(section);
             for (int i = 0; i < lines; i++)
@@ -574,9 +574,12 @@ namespace IL2DCE.MissionObjectModel
                 string value;
                 fileSrc.get(section, i, out key, out value);
                 fileSrc.delete(key);
-                foreach (var item in SubSectionAddKey)
+                if (SubSectionAddKey != null)
                 {
-                    fileSrc.delete(string.Format("{0}_{1}", key, item));
+                    foreach (var item in SubSectionAddKey)
+                    {
+                        fileSrc.delete(string.Format("{0}_{1}", key, item));
+                    }
                 }
             }
             fileSrc.delete(section);
